# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reconciliation & Feed Processing API",
    "description": "Production-ready FastAPI REST API backend wrapping feed generation, structural gating, tiered rule engine reconciliation, and diagnostics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/recon/run": {
      "post": {
        "tags": [
          "Reconciliation Engine"
        ],
        "summary": "Run Reconciliation Engine",
        "description": "Runs the tiered waterfall rule engine against the cache and sequential ingestion batches. Validates structural integrity, applies tiers (Exact, Date Tolerance, Reference Match, Amount Tolerance), and produces matched, remaining cache, unmatched exceptions, and ambiguous reviews.",
        "operationId": "run_reconciliation_api_recon_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recon/matches": {
      "get": {
        "tags": [
          "Reconciliation Engine"
        ],
        "summary": "Query Matched Transactions",
        "description": "Returns paginated matched transactions with optional account and tier filtering.",
        "operationId": "get_matches_api_recon_matches_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "account",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by account number/substring",
              "title": "Account"
            },
            "description": "Filter by account number/substring"
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by match tier (e.g. TIER_1_EXACT)",
              "title": "Tier"
            },
            "description": "Filter by match tier (e.g. TIER_1_EXACT)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recon/unmatched-cache": {
      "get": {
        "tags": [
          "Reconciliation Engine"
        ],
        "summary": "Query Remaining Unmatched Cache",
        "description": "Returns paginated internal cashbook/GL rows that were never matched by ingestion files.",
        "operationId": "get_unmatched_cache_api_recon_unmatched_cache_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "account",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recon/unmatched-ingest": {
      "get": {
        "tags": [
          "Reconciliation Engine"
        ],
        "summary": "Query Unmatched Ingestion Exceptions",
        "description": "Returns paginated bank statement transactions that could not find a match in the internal cache.",
        "operationId": "get_unmatched_ingest_api_recon_unmatched_ingest_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "account",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recon/ambiguous": {
      "get": {
        "tags": [
          "Reconciliation Engine"
        ],
        "summary": "Query Ambiguous Multi-Candidate Matches",
        "description": "Returns paginated records where multiple candidates competed, detailing tie-breaker scores and decisions.",
        "operationId": "get_ambiguous_matches_api_recon_ambiguous_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recon/download/{filename}": {
      "get": {
        "tags": [
          "Reconciliation Engine"
        ],
        "summary": "Download Result CSV",
        "description": "Download any reconciliation output or manifest CSV file directly.",
        "operationId": "download_csv_api_recon_download__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchConfigSchema": {
        "properties": {
          "date_tolerance_days": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Date Tolerance Days",
            "description": "Tier 2 date window fallback in days",
            "default": 3
          },
          "amount_abs_tolerance": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount Abs Tolerance",
            "description": "Tier 4 absolute currency slack",
            "default": 1.0
          },
          "amount_pct_tolerance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Amount Pct Tolerance",
            "description": "Tier 4 percentage slack on top of absolute",
            "default": 0.0
          },
          "direction_mode": {
            "type": "string",
            "enum": [
              "ignore",
              "same",
              "opposite"
            ],
            "title": "Direction Mode",
            "description": "'ignore' = sign convention unknown, 'same' = DR matches DR, 'opposite' = DR matches CR",
            "default": "ignore"
          }
        },
        "type": "object",
        "title": "MatchConfigSchema"
      },
      "PaginatedQueryResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "total_pages",
          "items"
        ],
        "title": "PaginatedQueryResponse"
      },
      "ReconRunRequest": {
        "properties": {
          "cache_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Path",
            "description": "Path to cache_gl_cashbook.csv. Defaults to config."
          },
          "manifest_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Path",
            "description": "Path to manifest.csv. Defaults to config."
          },
          "batches_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batches Dir",
            "description": "Directory containing batch CSVs. Defaults to config."
          },
          "out_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out Dir",
            "description": "Output directory for recon results. Defaults to config."
          },
          "config": {
            "$ref": "#/components/schemas/MatchConfigSchema",
            "description": "Tolerance and direction knobs"
          },
          "single_batch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Single Batch",
            "description": "Filename (e.g. ingest_batch_0001.csv) to test just one batch instead of full run."
          }
        },
        "type": "object",
        "title": "ReconRunRequest"
      },
      "ReconRunResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "total_matched": {
            "type": "integer",
            "title": "Total Matched"
          },
          "tier_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Tier Breakdown"
          },
          "cache_remaining_count": {
            "type": "integer",
            "title": "Cache Remaining Count"
          },
          "ingest_unmatched_count": {
            "type": "integer",
            "title": "Ingest Unmatched Count"
          },
          "ambiguous_count": {
            "type": "integer",
            "title": "Ambiguous Count"
          },
          "out_dir": {
            "type": "string",
            "title": "Out Dir"
          },
          "output_files": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Output Files"
          }
        },
        "type": "object",
        "required": [
          "message",
          "total_matched",
          "tier_breakdown",
          "cache_remaining_count",
          "ingest_unmatched_count",
          "ambiguous_count",
          "out_dir",
          "output_files"
        ],
        "title": "ReconRunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}